[Add] 리뷰적용
타입표시, socket keepAlive적용, myPage body 분리, getUserWithStatus에서 빈 배열로 나오는 결과 처리
