+++ client/src/composition/ChatRooms/ChatRoom.tsx
@@ -14,7 +14,7 @@ import Loader from 'components/Loader';
 import { useRef } from 'react';
 import useIntersect from 'hooks/useIntersectObserver';
 import { objToDate, dateToISO } from 'utils/dateUtil';
-import { DEFAULT } from 'constant';
+import { DEFAULT } from 'Constants';
 
 const Container = styled.div`
   width: 20rem;
