+++ client/src/composition/Feed/Feed.tsx
@@ -5,7 +5,7 @@ import FeedBody from './FeedBody';
 import FeedFooter from './FeedFooter';
 import CommentContainer from './FeedComment';
 import { IFeed } from 'react-components.d';
-import { DEFAULT } from 'constant';
+import { DEFAULT } from 'Constants';
 
 const FeedDiv = styled.div<{ feedSize: string }>`
   ${props => props.theme.borders.feedBorder};
