+++ client/src/composition/Feed/FeedComment/CommentPresentor.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import Profile from 'components/Profile';
 import { Comment } from 'react-components.d';
-import { DEFAULT } from 'constant';
+import { DEFAULT } from 'Constants';
 
 const CommentLine = styled.div`
   margin: 0.25rem 0;
