+++ client/src/composition/Feed/FeedComment/WriteCommentPresentor.tsx
@@ -9,7 +9,7 @@ import {
 import useInput, { IUseInput } from 'hooks/useInput';
 import Profile from 'components/Profile';
 import Button from 'components/Button';
-import { DEFAULT } from 'constant';
+import { DEFAULT } from 'Constants';
 
 const CommentForm = styled.div`
   position: relative;
