+++ client/src/composition/Header/index.tsx
@@ -9,7 +9,7 @@ import { useLazyQuery } from '@apollo/react-hooks';
 import { LOGOUT } from 'cache/client.gql';
 import { HeaderTabProvider } from 'stores/HeaderTabContext';
 import { HeaderAlarmCountProvider } from 'stores/HeaderTabCountContext';
-import { PAGE_PATHS, DEFAULT } from 'constant';
+import { PAGE_PATHS, DEFAULT } from 'Constants';
 import { Link } from 'react-router-dom';
 
 const HeaderWrapper = styled.div`
