+++ server/src/utils/pubsub.ts
@@ -5,6 +5,7 @@ export const pubsub = new RedisPubSub({
   connection: {
     host: '127.0.0.1',
     port: 6379,
+    keepAlive: 10000,
     retry_strategy: options => {
       // reconnect after
       return Math.max(options.attempt * 100, 3000);
